+++ src/api/viewsets/company/view.py
@@ -1,11 +1,14 @@
+from django.conf import settings
 from rest_framework import viewsets, status
 from rest_framework.response import Response
 
 from api.viewsets.company.model import Company
 from api.viewsets.company.serializer import CompanySerializer
 from api.viewsets.company.service import CompanyService
+from core.services.email import EmailService
+from core.utils.file import replace_keys
 from users.viewsets.user.permissions import IsAllowed
-from users.viewsets.user.service import UserService
+from users.viewsets.user.service import RegisterUserService
 
 
 class CompanyViewSet(viewsets.ModelViewSet):
@@ -21,5 +24,9 @@ class CompanyViewSet(viewsets.ModelViewSet):
     @staticmethod
     def create(request, *args, **kwargs):
         CompanyService.create_company(request.data['company'])
-        UserService.create_user(request.data['user'])
-        return Response({}, status=status.HTTP_201_CREATED)
+        token = RegisterUserService().create_user(request.data['user'])
+        email_service = EmailService()
+        body = email_service.get_template(f"{settings.BASE_DIR}/templates/account_verification.html")
+        body = replace_keys(body, '##VERIFICATION_TOKEN##', token)
+        return email_service.send_sendgrid_email(receiver_email=request.data['user']['email'],
+                                                 subject='youML - Account verification', body=body)