+++ src/users/viewsets/user/view.py
@@ -1,10 +1,13 @@
 from rest_framework import viewsets, status
 from rest_framework.response import Response
 
+from core.services.email import EmailService
+from core.utils.file import replace_keys
 from users.models import User
 from users.viewsets.user.permissions import IsAllowed
 from users.viewsets.user.serializer import UserSerializer
-from users.viewsets.user.service import UserService
+from users.viewsets.user.service import RegisterUserService
+from django.conf import settings
 
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -19,5 +22,9 @@ class UserViewSet(viewsets.ModelViewSet):
 
     @staticmethod
     def create(request, *args, **kwargs):
-        UserService.create_user(request.data)
-        return Response({}, status=status.HTTP_201_CREATED)
+        token = RegisterUserService().create_user(request.data)
+        email_service = EmailService()
+        body = email_service.get_template(f"{settings.BASE_DIR}/templates/account_verification.html")
+        body = replace_keys(body, '##VERIFICATION_TOKEN##', token)
+        return email_service.send_sendgrid_email(receiver_email=request.data['email'],
+                                                 subject='youML - Account verification', body=body)